requirements.txt
alembic==1.12.0
blinker==1.6.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
environs==9.5.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
logging-formatter-anticrlf==1.2.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
numpy==1.26.0
packaging==23.1
pandas==2.1.1
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
tzdata==2023.3
waitress==2.1.2
Werkzeug==2.3.7
